Make the episodes and battle levels list loading process be threaded
