+++ changelog.txt
@@ -1,5 +1,8 @@
 TheXTech Changelog.
 
+Changes for 1.3.5.1
+-Episodes and battle levels lists now loading in the separated thread with a progress showing
+
 Changes for 1.3.5
 -Fixed the gamepad control stuck because of axes returned value that doesn't equal the initial state on some devices
 -Fixed the invalid conversion of old sprites with a mask that is smaller than the front
